* chore: update files for new gitflow (#227)

* chore: update files for new-release

* chore: update files for new gitflow

* chore: update readme (#229)

* docs: update tutorials

* docs: update tutorials (#230)

* feat: new datasets and docs (#231)

* feat: new datasets and docs

* fix: fmt

* docs: rename lib (#234)

* fix: datasets labels (#245)

* feature: update dataset_shift module (#248)

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* fix: update dataset_shift code and plots

* fix: update dataset_shift code and plots

* fix: update dataset_shift code and plots

* fix: percent_degradation to ADP nomenclature

* chore(deps): bump actions/checkout from 4.1.7 to 4.2.2 (#250)

Bumps [actions/checkout](https://github.com/actions/checkout) from 4.1.7 to 4.2.2.
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/v4.1.7...v4.2.2)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(deps): bump actions/setup-python from 5.2.0 to 5.3.0 (#251)

Bumps [actions/setup-python](https://github.com/actions/setup-python) from 5.2.0 to 5.3.0.
- [Release notes](https://github.com/actions/setup-python/releases)
- [Commits](https://github.com/actions/setup-python/compare/v5.2.0...v5.3.0)

---
updated-dependencies:
- dependency-name: actions/setup-python
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* feat: New Robustness Regression Attacker: Ridge Poissoner (#241)

* feat: adding adaptive and randflip initializers for regression poisoners

* feat: adding ridge-based regression poisoner

* test: adding unit tests for ridge-based regression poisoner

* chore: updating tutorial with ridge-based poisoner

* docs: updating regression poisoner docs with ridge regressor

* docs: adding docstrings for ridge regressor poisoner

* chore: fixing lint

* chore: updating docstrings for poisoners

* chore: updloading notebook file for regression poisoner tutorial

* docs: updating possible parameters in docstrings

* feat: new methods for measuring security privacy risk score (#254)

* feat: adding source code and tutorial for privacy risk score

* tests: adding unit test for privacy risk score metric

* docs: updating reference docstrings for privacy risk score metric

* docs: updating docs for example gallery for security module

* feat: adding tutorial for privacy risk score with different models

* chore: adding tutorial for example gallery

* chore: adding tutorial notebook files for docs section

* Update security.rst

---------


* feat: new xai metrics based on globa feature importance, local feature importance and surrogate  (#255)

* internal changes

* fix: pfi regression and shap

* update: shap and contrast

* chore: improved xai metrics

* fix: paper regression results

* ok

* chore:  xai-metrics updated

* chore: format files

* test: add test for new metrics

* docs: update documentation and docstring format

* chore: hatch dmt

---------


* chore: Improve Accuracy Degradation Profile Documentation  (#249)

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* feature: update dataset_shift module

* fix: update dataset_shift code and plots

* fix: update dataset_shift code and plots

* fix: update dataset_shift code and plots

* fix: percent_degradation to ADP nomenclature

* fix: uptade documentation

* chore: clean adp tutorial

* chore: remove inccorect reference

* chore: format

* doc: fix docstring

---------


* docs:update tutorials (#257)

* docs: updating docstrings for regression poissoners (#256)

* fix: load dataset (#259)

* Delete paper directory (#262)

* Delete README.rst

* Delete workflows directory (#263)

* Features/improvementes (#265)

* chore: push updates

* fix: compas and add description

* fix: format

---------


* add: bias tradeoff example (#266)

* add: bank marketing description (#267)

* chore: updating hackaton notebook with privacy risk score interpretation (#271)

* docs: update documetnation (#274)

* fix: merge conflicts

---------

+++ src/holisticai/explainability/metrics/global_feature_importance/_rank_alignment.py
@@ -29,7 +29,7 @@ class RankAlignment:
                 / len(set(top_feature_names).union(top_cond_feature_names))
             )
         if aggregation:
-            return np.mean(similarities)
+            return float(np.mean(similarities))
         return similarities
